+++ tests/testthat/test-atlantis-calibrate.R
@@ -15,6 +15,13 @@ res_full <- mod |>
   generate_calibration_table(
     path_to_mock("right_full.yaml")
   )
+res_relative <- mod |>
+  generate_calibration_table(
+    path_to_mock("right_relative.yaml")
+  )
+# issues with last digits
+res_relative$min <- round(res_relative$min, 8)
+res_relative$max <- round(res_relative$max, 8)
 
 test_that("Error handling works correctly", {
   expect_error(
@@ -45,6 +52,17 @@ test_that("Error handling works correctly", {
       ),
     "Group info required"
   )
+  #
+  expect_error(
+    new_atlantis() |>
+      atlantis_load_files(
+        atlantis_examples("inputs", "tiny_groups.csv")
+      ) |>
+      generate_calibration_table(
+        path_to_mock("right_relative.yaml")
+      ),
+    "Cannot find value for"
+  )
 })
 
 
@@ -53,6 +71,7 @@ test_that("Works as expected", {
     res_simplest,
     data.frame(
       name = "rec_m",
+      cur_value = 1,
       min = -Inf,
       max = Inf,
       position = 1L,
@@ -70,6 +89,26 @@ test_that("Works as expected", {
         rep("mum_YPH", 2),
         rep("mum_RDG", 2),
         "rec_m"
+      ),
+      cur_value = c(
+        0.036549944,
+        0.003063438,
+        0.00167827,
+        0.00071767,
+        0.0004755,
+        0.00606736,
+        0.003030164,
+        0.001290054,
+        0.001206866,
+        0.000960844,
+        0.000836366,
+        0.000447656,
+        0.000447656,
+        0.006924034,
+        0.0017007,
+        0.009661128,
+        0.004078176,
+        1
       ),
       min = c(rep(-6, 13), rep(-3, 4), -Inf),
       max = c(rep(-1, 17), Inf),
@@ -95,6 +134,27 @@ test_that("Works as expected", {
       ),
       transf = c(rep("pow10", 17), "identity"),
       source_file = rep("biology_prm", 18)
+    )
+  )
+})
+
+test_that("Works as expected with is_factor set to true", {
+  withr::local_options(list(digits = 6))
+  expect_identical(
+    res_relative,
+    data.frame(
+      name = c("mum_GZS", "mum_GZS", "mum_GZS", "mum_WAE"),
+      cur_value = c(0.036549944, 0.003063438, 0.00167827, 0.00606736),
+      min = c(-Inf, -Inf, -3.77513817, 0.00303368),
+      max = c(Inf, Inf, -1.77513817, 0.00910104),
+      position = c(1L, 2L, 3L, 1L),
+      transf = c("pow10", "pow10", "pow10", "identity"),
+      source_file = c(
+        "biology_prm",
+        "biology_prm",
+        "biology_prm",
+        "biology_prm"
+      )
     )
   )
 })